Revision IDs need to be long everywhere
In df21c73a5636676cff99eeb2a1c7a1e2218f88bf, revision IDs have been changed
to 64-bit longs, which is now necessary on Wikidata. However, wrong explicit
`(int)` casts remained in the code at multiple places, causing exceptions.

Fixes #112

+++ WikiClientLibrary/Generators/Primitive/WikiPageGenerator.cs
@@ -137,7 +137,7 @@ public abstract class WikiPageGenerator : WikiPageGenerator<WikiPageStub>
     /// <inheritdoc />
     protected override WikiPageStub ItemFromJson(JToken json)
     {
-        return new WikiPageStub((int)json["pageid"], (string)json["title"], (int)json["ns"]);
+        return new WikiPageStub((long)json["pageid"], (string)json["title"], (int)json["ns"]);
     }
 
 }
